Type the callAPI function and actions
It now correctly returns a Thunk which can be dispatched to get a
promise with a typed response. No more Thunk<any>/Thunk<Promise<any>>
bs.

The way we handle Action and PromiseAction is awful. The PromiseAction
is handled by the middleware, while the Action is returned by the
callAPI, for some reason.

Seems to work fine, but I'm getting a ts error when dispatching the
action in the call api function. Would appreciate some help here.